Move world generation utilities into engine
