+++ src/app/api/map/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from 'next/server';
-import { createWorldGenerator } from '@/lib/world/generator';
+import { createWorldGenerator } from '@engine';
 
 export async function GET(req: NextRequest) {
   const { searchParams } = new URL(req.url);
